requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# AI and Content Generation
openai==1.6.1
langchain==0.1.0
tiktoken==0.5.2

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.16.0
scrapy==2.11.0

# Image Generation
pillow==10.1.0
stability-sdk==0.8.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Scheduling
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# SEO and Content Analysis
yake==0.4.8
textstat==0.7.3
newspaper3k==0.2.8

# Web Dashboard
streamlit==1.29.0
plotly==5.18.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
httpx==0.25.2
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging and Monitoring
loguru==0.7.2
sentry-sdk==1.38.0 